refactor: move `installDepsFromPackageJson` function to JsPackageManager class and subclasses
